moved global to elife-ui
