+++ packages/component-elife-ui/client/global/AppBar/AppBar.test.js
