+++ packages/component-elife-ui/client/global/CookieNotice/helpers.js
