+++ packages/component-elife-ui/client/global/CookieNotice/helpers.test.js
