+++ packages/component-elife-ui/client/global/CookieNotice/index.test.js
