+++ packages/component-elife-ui/client/global/RedirectLayout/index.js
