+++ packages/component-elife-ui/client/global/layout/TwoColumnLayout.md
