+++ packages/component-elife-ui/config/jest.config.client.js
@@ -6,7 +6,7 @@ module.exports = {
   transformIgnorePatterns: ['/node_modules/(?!@?pubsweet|xpub)'],
   moduleNameMapper: {
     '\\.s?css$': 'identity-obj-proxy',
-    '^ui(.*)$': '<rootDir>/../../packages/component-elife-ui/client/ui$1',
-    '^global(.*)$': '<rootDir>/../../app/components/global$1',
+    '^ui(.*)$': '<rootDir>/../client/ui$1',
+    '^global(.*)$': '<rootDir>/../client/global$1',
   },
 }
