+++ packages/component-elife-ui/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@elifesciences/component-elife-ui",
+  "version": "1.0.0",
+  "main": "index.js",
+  "license": "MIT"
+}
