+++ webpack/webpack.styleguide.config.js
@@ -25,7 +25,10 @@ module.exports = {
     alias: {
       config: path.join(__dirname, '../styleguide/config.json'),
       ui: path.resolve(__dirname, '../packages/component-elife-ui/client/ui'),
-      global: path.resolve(__dirname, '../app/components/global'),
+      global: path.resolve(
+        __dirname,
+        '../packages/component-elife-ui/client/global',
+      ),
     },
   },
   plugins: [new CopyWebpackPlugin([{ from: './assets', to: 'assets' }])],
